backend/requirements.txt
fastapi
uvicorn[standard] 
python-dotenv
psycopg2
supabase
asyncpg
python-jose
httpx
python-multipart